Add BaseFrameLikeShapeUtil abstract class
Extract common frame-like shape behavior into a reusable base class
that extends BaseBoxShapeUtil. This consolidates isFrameLike,
providesBackgroundForChildren, canReceiveNewChildrenOfType, getClipPath,
onDragShapesIn, and onDragShapesOut into one place so custom frame-like
shapes don't need to reimplement each method individually.

FrameShapeUtil now extends BaseFrameLikeShapeUtil, removing ~80 lines
of boilerplate.
