+++ packages/tldraw/src/lib/shapes/frame/FrameShapeUtil.tsx
@@ -1,5 +1,5 @@
 import {
-	BaseBoxShapeUtil,
+	BaseFrameLikeShapeUtil,
 	DefaultColorStyle,
 	Geometry2d,
 	Group2d,
@@ -7,13 +7,9 @@ import {
 	SVGContainer,
 	SvgExportContext,
 	TLClickEventInfo,
-	TLDragShapesOutInfo,
-	TLDragShapesOverInfo,
 	TLEditStartInfo,
 	TLFrameShape,
 	TLFrameShapeProps,
-	TLResizeInfo,
-	TLShape,
 	TLShapePartial,
 	TLShapeUtilConstructor,
 	clamp,
@@ -22,7 +18,6 @@ import {
 	frameShapeProps,
 	getColorValue,
 	lerp,
-	resizeBox,
 	toDomPrecision,
 	useColorMode,
 	useValue,
@@ -79,7 +74,7 @@ export interface FrameShapeOptions extends ShapeOptionsWithDisplayValues<
 }
 
 /** @public */
-export class FrameShapeUtil extends BaseBoxShapeUtil<TLFrameShape> {
+export class FrameShapeUtil extends BaseFrameLikeShapeUtil<TLFrameShape> {
 	static override type = 'frame' as const
 	static override props = frameShapeProps
 	static override migrations = frameShapeMigrations
@@ -362,22 +357,6 @@ export class FrameShapeUtil extends BaseBoxShapeUtil<TLFrameShape> {
 		return path
 	}
 
-	override providesBackgroundForChildren(): boolean {
-		return true
-	}
-
-	override getClipPath(shape: TLFrameShape) {
-		return this.editor.getShapeGeometry(shape.id).vertices
-	}
-
-	override canReceiveNewChildrenOfType(shape: TLShape) {
-		return !shape.isLocked
-	}
-
-	override onResize(shape: any, info: TLResizeInfo<any>) {
-		return resizeBox(shape, info)
-	}
-
 	override getInterpolatedProps(
 		startShape: TLFrameShape,
 		endShape: TLFrameShape,
@@ -435,64 +414,6 @@ export class FrameShapeUtil extends BaseBoxShapeUtil<TLFrameShape> {
 		return {
 			id: shape.id,
 			type: shape.type,
-		}
-	}
-
-	override onDragShapesIn(
-		shape: TLFrameShape,
-		draggingShapes: TLShape[],
-		{ initialParentIds, initialIndices }: TLDragShapesOverInfo
-	) {
-		const { editor } = this
-
-		if (draggingShapes.every((s) => s.parentId === shape.id)) return
-
-		// Check to see whether any of the shapes can have their old index restored
-		let canRestoreOriginalIndices = false
-		const previousChildren = draggingShapes.filter((s) => shape.id === initialParentIds.get(s.id))
-
-		if (previousChildren.length > 0) {
-			const currentChildren = compact(
-				editor.getSortedChildIdsForParent(shape).map((id) => editor.getShape(id))
-			)
-			if (previousChildren.every((s) => !currentChildren.find((c) => c.index === s.index))) {
-				canRestoreOriginalIndices = true
-			}
-		}
-
-		// I can't imagine this happening, but if any of the children are the ancestor of the frame, quit here
-		if (draggingShapes.some((s) => editor.hasAncestor(shape, s.id))) return
-
-		// Reparent the shapes to the new parent
-		editor.reparentShapes(draggingShapes, shape.id)
-
-		// If we can restore the original indices, then do so
-		if (canRestoreOriginalIndices) {
-			for (const shape of previousChildren) {
-				editor.updateShape({
-					id: shape.id,
-					type: shape.type,
-					index: initialIndices.get(shape.id),
-				})
-			}
-		}
-	}
-
-	override onDragShapesOut(
-		shape: TLFrameShape,
-		draggingShapes: TLShape[],
-		info: TLDragShapesOutInfo
-	): void {
-		const { editor } = this
-		// When a user drags shapes out of a frame, and if we're not dragging into a new shape, then reparent
-		// the dragging shapes (that are current children of the frame) onto the current page instead
-		if (!info.nextDraggingOverShapeId) {
-			editor.reparentShapes(
-				draggingShapes.filter(
-					(s) => s.parentId === shape.id && this.canReceiveNewChildrenOfType(s)
-				),
-				editor.getCurrentPageId()
-			)
 		}
 	}
 }
